requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Werkzeug
Pillow
imutils
face_recognition
opencv-python
numpy
psycopg2-binary


Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Login==0.6.3
Werkzeug==3.1.3
Pillow==11.3.0
imutils==0.5.4
face_recognition==1.3.0
opencv-python==4.10.0.84
numpy==2.3.2
psycopg2-binary==2.9.10